pyproject.toml
[project]
name = "gs_utils"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "viser>=1.0.24",
    "imageio[ffmpeg]>=2.37.3",
    "scikit-learn>=1.8.0",
    "tqdm>=4.67.3",
    "torchmetrics[image]>=1.9.0",
    "opencv-python>=4.11.0.86",
    "tyro>=0.8.8",
    "pillow>=12.1.1",
    "piexif>=1.1.3",
    "tensorboard>=2.20.0",
    "tensorly>=0.9.0",
    "pyyaml>=6.0.3",
    "matplotlib>=3.10.8",
    "splines>=0.3.3",
    "nerfacc>=0.5.3,<0.6",
    "pycolmap>=4.0.2,<5",
    "nerfview",
    "fused-ssim",
    "fused-bilagrid",
    "ppisp==1.0.0",
    "piexif",
    "pydantic>=2.12.5,<3",
    "jaxtyping>=0.3.9,<0.4",
    "gsplat>=1.5.3,<2",
]

[tool.uv.sources]
nerfview = { git = "https://github.com/nerfstudio-project/nerfview", rev = "4538024fe0d15fd1a0e4d760f3695fc44ca72787" }
fused-ssim = { git = "https://github.com/rahul-goel/fused-ssim" }
fused-bilagrid = { git = "https://github.com/harry7557558/fused-bilagrid" }
ppisp = { git = "https://github.com/nv-tlabs/ppisp" }

[tool.uv.extra-build-dependencies]
fused-ssim = ["torch", "setuptools"]
fused-bilagrid = ["torch", "setuptools"]
nerfview = ["torch", "setuptools"]
ppisp = ["torch", "setuptools"]


# --- Formatting ---

[tool.ruff]
line-length = 80  # Must match above


[tool.ruff.lint] # Mostly just copied over from mojo
ignore = [
  "ANN002",     # TODO: Missing type annotation for *args
  "ANN003",     # TODO: Missing type annotation for **kwargs
  "ANN204",     # TODO: Missing return type annotation for special method
  "ANN401",     # TODO: Dynamically typed expressions (typing.Any) are disallowed in `**kwargs`
  "B008",       # Do not perform function call in argument defaults
  "B017",       # Do not assert blind exception: `Exception`
  "B028",       # No explicit `stacklevel` keyword argument found
  "D105",       # undocumented-magic-method
  "D107",       # Missing docstring in __init__ (constructor args go in class docstring)
  "E402",       # TODO: import top of file
  "E722",       # bare except
  "E731",       # TODO: Don't assign lambda
  "E741",       # TODO: Ambiguous variable name
  "F403",       # glob imports
  "F405",       # TODO: glob imports undefined
  "F821",       # TODO: Undefined names, means script is invalid
  "F841",       # TODO: unused variables
  "FURB122",    # Questionable readability improvement
  "RUF005",     # Consider iterable unpacking instead of concatenation
  "RUF010",     # Use explicit conversion flag
  "RUF012",     # Mutable class attributes should be annotated with `typing.ClassVar`
  "RUF015",     # Prefer `next(iter(iterable))` over single element slice
  "SIM102",     # Collapsible if
  "SIM105",     # Suppressible exception
  "SIM108",     # Use ternary operator
  "SIM114",     # Combine arms with or
  "SIM115",     # Use context manager for opening files
  "SIM117",     # Merge with statements
  "SIM300",     # YODA conditions
  "SIM905",     # Use a list literal instead of str.split
  "F722",       # Unexpected token in forward annotation, clashes with jaxtyping
]
extend-select = [
  "ANN",        # Require function type annotations
  "B",          # Bugbear lints
  "D",          # Pydocstyle lints
  "FURB",       # Refurb lints for modernization
  "I",          # Import sorting lints
  "RUF",        # Ruff specific lints
  "SIM",        # Simplify lints
  "T10",        # Checks for the presence of debugger calls and imports.
  "UP",         # Type annotations should not use Union, Optional, etc
  "ISC",        # Good use of string concatenation
  "SIM",        # Common simplification rules
  "TID",        # Some good import practices
  "PTH",        # Use pathlib instead of os.path
  "NPY",        # Some numpy-specific things

]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ANN", "D"] # Docstrings aren't strictly required in test files
